refactor(logic): Slice B world basename collision purge/port (Refs #4090)
Delete covered tile_control orphan; port unique province_lookup and
minor_military_parity asserts into colonizethis_world, then remove the
logic twins. Remaining connectivity/naval/movement orphans deferred.
